+++ jest.config.js
@@ -1,0 +1,10 @@
+// jest.config.js
+module.exports = {
+    moduleFileExtensions: ['js', 'json', 'ts'],
+    rootDir: 'src',
+    testRegex: '.spec.ts$',
+    transform: {
+        '^.+\\.(t|j)s$': 'ts-jest'
+    },
+    coverageDirectory: '../coverage'
+};