+++ src/constants.ts
@@ -1,0 +1,5 @@
+/**
+ * @module netjs-console
+ */
+
+export const COMMANDER_SERVICE_TOKEN = 'Commander';