+++ src/decorators.ts
@@ -1,0 +1,10 @@
+/**
+ * @module nestjs-console
+ */
+
+import { Inject } from '@nestjs/common';
+import { COMMANDER_SERVICE_TOKEN } from './constants';
+
+export const InjectCommander = () => {
+    return Inject(COMMANDER_SERVICE_TOKEN);
+};