+++ src/index.ts
@@ -1,3 +1,6 @@
+export { Command } from './commander';
+export * from './console';
+export * from './constants';
+export * from './decorators';
 export * from './module';
 export * from './service';
-export * from './console';