+++ src/service.ts
@@ -1,34 +1,36 @@
-import { Injectable, Inject } from '@nestjs/common';
-import { CommanderStatic, Command } from 'commander';
+import { Injectable } from '@nestjs/common';
 import ora from 'ora';
-
-export interface PatchedCommander extends Command {
-    forwardSubCommand: () => CommanderStatic;
-}
+import { Command, forwardSubCommands } from './commander';
+import { InjectCommander } from './decorators';
 
 @Injectable()
 export class ConsoleService {
-    private readonly loader = ora();
+    constructor(@InjectCommander() protected readonly cli: Command) {}
 
-    constructor(@Inject('Commander') private readonly cli: CommanderStatic) {}
+    static createSpinner(text?: string) {
+        return ora(text);
+    }
 
     getCli() {
         return this.cli;
     }
 
-    log(...args: any[]) {
-        console.log(...args);
-    }
-
-    createSpinner(text?: string): any {
-        return ora(text);
-    }
-
     init(argv: string[]) {
-        this.cli.on('command:*', () => this.cli.help());
-        this.cli.parse(argv);
-        if (argv.length === 2) {
+        this.cli.on('command:*', () => {
             this.cli.help();
+        });
+        if (argv.length === 2) {
+            argv.push('');
         }
+        this.cli.parse(argv);
+    }
+
+    subCommands(
+        parent: Command,
+        command: string,
+        description: string
+    ): Command {
+        const subCommand = parent.command(command).description(description);
+        return forwardSubCommands.bind(subCommand)();
     }
 }